refactor: separate utils and code coverage information (#174)
@W-9061799@
